+++ examples/simplecqrs/simplecqrs/inventoryitem.go
@@ -31,7 +31,8 @@ func (a *InventoryItem) Create(name string) error {
 	}
 
 	a.Apply(ycq.NewEventMessage(a.AggregateID(),
-		&InventoryItemCreated{ID: a.AggregateID(), Name: name}), true)
+		&InventoryItemCreated{ID: a.AggregateID(), Name: name},
+		ycq.Int(a.CurrentVersion())), true)
 
 	return nil
 }
@@ -43,7 +44,8 @@ func (a *InventoryItem) ChangeName(newName string) error {
 	}
 
 	a.Apply(ycq.NewEventMessage(a.AggregateID(),
-		&InventoryItemRenamed{ID: a.AggregateID(), NewName: newName}), true)
+		&InventoryItemRenamed{ID: a.AggregateID(), NewName: newName},
+		ycq.Int(a.CurrentVersion())), true)
 
 	return nil
 }
@@ -59,7 +61,8 @@ func (a *InventoryItem) Remove(count int) error {
 	}
 
 	a.Apply(ycq.NewEventMessage(a.AggregateID(),
-		&ItemsRemovedFromInventory{ID: a.AggregateID(), Count: count}), true)
+		&ItemsRemovedFromInventory{ID: a.AggregateID(), Count: count},
+		ycq.Int(a.CurrentVersion())), true)
 
 	return nil
 }
@@ -71,7 +74,8 @@ func (a *InventoryItem) CheckIn(count int) error {
 	}
 
 	a.Apply(ycq.NewEventMessage(a.AggregateID(),
-		&ItemsCheckedIntoInventory{ID: a.AggregateID(), Count: count}), true)
+		&ItemsCheckedIntoInventory{ID: a.AggregateID(), Count: count},
+		ycq.Int(a.CurrentVersion())), true)
 
 	return nil
 }
@@ -83,7 +87,8 @@ func (a *InventoryItem) Deactivate() error {
 	}
 
 	a.Apply(ycq.NewEventMessage(a.AggregateID(),
-		&InventoryItemDeactivated{ID: a.AggregateID()}), true)
+		&InventoryItemDeactivated{ID: a.AggregateID()},
+		ycq.Int(a.CurrentVersion())), true)
 
 	return nil
 }
